==> Code/0_tif_to_geojson.py <==
# -*- coding: utf-8 -*-
"""0_tif to geojson.ipynb

Automatically generated by Colaboratory.

Original file is located at
    https://colab.research.google.com/drive/1k6uOZ46Y7oQmo3BsPsCSn3SN80D8K6dv
"""

from google.colab import drive #to retrieve data from drive 
drive.mount('/content/drive')

cd 'drive/My Drive/Data/OSM'

!pip install rasterio

import rasterio
import rasterio.features
import rasterio.warp
from csv import writer
import pandas as pd
import numpy as np

district = ['Ranchi','Allahabad', 'Nanded', 'Patna', 'Raipur', 'Chittaurgarh' ]
district.pop(0)
# dist_coords = pd.read_csv("district_coordinates.csv")
# print(dist_coords)
# print()
# bboxes = dict()
# for i, j in dist_coords.iterrows():
#   bboxes[j['District_Name'].lower()] = (j['MinLat'], j['MaxLat'], j['MinLong'], j['MaxLong'])
# print(bboxes)
# min_lon=float(dist_coords.loc[dist_coords['District_Name'] == 'Dhanbad']['MinLong'])
# min_lat=float(dist_coords.loc[dist_coords['District_Name'] == 'Dhanbad']['MinLat'])
# max_lon=float(dist_coords.loc[dist_coords['District_Name'] == 'Dhanbad']['MaxLong'])
# max_lat=float(dist_coords.loc[dist_coords['District_Name'] == 'Dhanbad']['MaxLat'])
# print([min_lat, max_lat, min_lon, max_lon])
# print(dist_coords.to_dict('list'))

with rasterio.open(district[0]+'/sentinel2a_'+district[0]+'_2019_year_median.tif') as dataset:

    # Read the dataset's valid data mask as a ndarray.
    mask = dataset.dataset_mask()

    # Extract feature shapes and values from the array.
    for geom, val in rasterio.features.shapes(
            mask, transform=dataset.transform):

        # Transform shapes from the dataset's own coordinate
        # reference system to CRS84 (EPSG:4326).
        geom = rasterio.warp.transform_geom(
            dataset.crs, 'EPSG:4326', geom, precision=6)

        # Print GeoJSON shapes to stdout.
        print(geom)

        # print(geom['coordinates'][0])
        boundary = geom['coordinates'][0]
        bbox = [district[0], boundary[1][1], boundary[0][1], boundary[0][0], boundary[2][0]]
        print(bbox)
        with open('district_coordinates.csv', 'a') as f_object:
  
          # Pass this file object to csv.writer()
          # and get a writer object
          writer_object = writer(f_object)
        
          # Pass the list as an argument into
          # the writerow()
          writer_object.writerow(bbox)
        
          #Close the file object
          f_object.close()

# Commented out IPython magic to ensure Python compatibility.
#path that contains folder you want to copy
# %cd /gdrive/My Drive/...
# %cp -av YOUR_FOLDER NEW_FOLDER_COPY

cd 'GEE_Data'
