==> Code/2_compressclasses_to_bu_nbu.py <==
# -*- coding: utf-8 -*-
"""2_compressClasses_to_BU_NBU.ipynb

Automatically generated by Colaboratory.

Original file is located at
    https://colab.research.google.com/drive/13wVOvcbiyBDEhpvcXQSCwm0YsVckMqeJ
"""



from google.colab import drive #to retrieve data from drive 
drive.mount('/content/drive/')

cd 'drive/My Drive/Data/OSM'

from PIL import Image
import numpy as np
import pandas as pd
import os, sys

# define color coding used in prediction images
Background = 0
Green = 1
Water = 2
Builtup = 3
Barrenland = 4

# districts=['Bangalore', 'Chennai', 'Delhi', 'Gurgaon', 'Hyderabad', 'Kolkata', 'Mumbai']
# years = ['2016', '2017', '2018', '2019']
districts=['Jamui', 'Bokaro', 'Hardoi']
years = ['2019', '2016']

destination_directory = 'BU_NBU_maps'
os.makedirs( destination_directory, exist_ok = True )

for district in districts:
    for year in years:
        image_filename = 'Landcover_Predictions_Using_IndiaSat/'+district+'/'+district+'_prediction_'+year+'.png'
        image = np.array( Image.open(image_filename) )
         
        # keep built-up pixels as it is i.e. BU with color code (1 * 65)
        image[ image == 3 ] = 65
        # keep green, water, and barrenland pixels as non-built-up i.e. NBU with color code (2 x 65)
        image[ image == 1 ] = 130
        image[ image == 2 ] = 130
        image[ image == 4 ] = 130

        os.makedirs( destination_directory+'/'+district, exist_ok = True )

        image = ( Image.fromarray(image) ).convert("L")
        image.save(destination_directory+'/'+district+'/'+district+'_BU_NBU_'+year+'.png')
    print("BU/NBU Map created for ",district)
    
print("\n#### Check ",destination_directory," directory for results ####\n")
